+++ lionagi/__init__.py
@@ -2,9 +2,10 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+from lionagi.operations import types as op
 from lionagi.operatives import types as ops_types
 from lionagi.protocols import types as types
-from lionagi.service.types import iModel
+from lionagi.service.imodel import iModel
 from lionagi.session.session import Branch, Session
 from lionagi.version import __version__
 
@@ -17,5 +18,6 @@ __all__ = (
     "LiteiModel",
     "types",
     "ops_types",
+    "op",
     "__version__",
 )
